refactor: removed community member doctype
