+++ community/community/doctype/community_event_volunteer/community_event_volunteer.json
@@ -13,13 +13,13 @@
    "fieldtype": "Link",
    "in_list_view": 1,
    "label": "Member",
-   "options": "Community Member"
+   "options": "User"
   }
  ],
  "index_web_pages_for_search": 1,
  "istable": 1,
  "links": [],
- "modified": "2021-02-15 12:03:31.153575",
+ "modified": "2021-05-21 12:15:51.286478",
  "modified_by": "Administrator",
  "module": "Community",
  "name": "Community Event Volunteer",
