--- community/community/doctype/community_member/community_member.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2021, Frappe and contributors
-// For license information, please see license.txt
-
-frappe.ui.form.on('Community Member', {
-	// refresh: function(frm) {
-
-	// }
-});
