--- community/community/doctype/community_member/community_member.json
@@ -1,155 +1,0 @@
-{
- "actions": [],
- "allow_guest_to_view": 1,
- "allow_rename": 1,
- "creation": "2021-02-12 15:47:23.591567",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "full_name",
-  "email",
-  "enabled",
-  "column_break_4",
-  "username",
-  "email_preference",
-  "section_break_7",
-  "bio",
-  "section_break_9",
-  "role",
-  "photo",
-  "column_break_12",
-  "short_intro",
-  "route",
-  "abbr"
- ],
- "fields": [
-  {
-   "default": "1",
-   "fieldname": "enabled",
-   "fieldtype": "Check",
-   "label": "Enabled"
-  },
-  {
-   "fieldname": "full_name",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "in_standard_filter": 1,
-   "label": "Full Name",
-   "reqd": 1
-  },
-  {
-   "allow_in_quick_entry": 1,
-   "fieldname": "role",
-   "fieldtype": "Select",
-   "label": "Role",
-   "options": "\nBoard\nDirector\nVolunteer\nSpeaker"
-  },
-  {
-   "allow_in_quick_entry": 1,
-   "fieldname": "photo",
-   "fieldtype": "Attach Image",
-   "label": "Photo"
-  },
-  {
-   "fieldname": "short_intro",
-   "fieldtype": "Data",
-   "label": "Short Intro"
-  },
-  {
-   "allow_in_quick_entry": 1,
-   "fieldname": "bio",
-   "fieldtype": "Markdown Editor",
-   "label": "Bio"
-  },
-  {
-   "fieldname": "route",
-   "fieldtype": "Data",
-   "label": "Route"
-  },
-  {
-   "fieldname": "email",
-   "fieldtype": "Data",
-   "in_standard_filter": 1,
-   "label": "Email",
-   "options": "Email",
-   "reqd": 1,
-   "unique": 1
-  },
-  {
-   "allow_in_quick_entry": 1,
-   "fieldname": "username",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "User Name",
-   "unique": 1
-  },
-  {
-   "fieldname": "email_preference",
-   "fieldtype": "Select",
-   "label": "Email preference",
-   "options": "Email on every Message\nOne Digest Mail per day\nNever"
-  },
-  {
-   "fieldname": "column_break_4",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "section_break_7",
-   "fieldtype": "Section Break"
-  },
-  {
-   "fieldname": "section_break_9",
-   "fieldtype": "Section Break"
-  },
-  {
-   "fieldname": "column_break_12",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "abbr",
-   "fieldtype": "Data",
-   "label": "Abbr",
-   "read_only": 1
-  }
- ],
- "index_web_pages_for_search": 1,
- "links": [],
- "modified": "2021-04-28 11:22:35.402217",
- "modified_by": "Administrator",
- "module": "Community",
- "name": "Community Member",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  },
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "Website Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "quick_entry": 1,
- "search_fields": "full_name",
- "sort_field": "modified",
- "sort_order": "DESC",
- "title_field": "full_name",
- "track_changes": 1
-}
