--- community/community/doctype/community_member/templates/community_member.html
@@ -1,25 +1,0 @@
-{% extends "templates/web.html" %}
-{% block page_content %}
-<div class="py-20 row">
-    {% if photo %}
-    <div class="col-sm-2 border border-dark">
-        <img src="{{ photo }}" alt="{{ full_name }}">
-    </div>
-    {% else %}
-    <div class="col-sm-2">
-        <div class="standard-image" style="font-size: 30px;">{{ abbr }}</div>
-    </div>
-    {% endif %}
-    <div class="col">
-        <h1>{{ full_name }}</h1>
-        {% if short_intro %}
-        <p class="lead"> {{ short_intro }} </p>
-        {% endif %}
-        {% if bio %}
-        <p class="markdown-style"> {{ frappe.utils.md_to_html(bio) }} </p>
-        {% endif %}
-    </div>
-</div>
-{% endblock %}
-
-<!-- this is a sample default web page template -->
