--- community/community/doctype/community_member/templates/community_member_row.html
@@ -1,4 +1,0 @@
-<div>
-	<a href="{{ doc.route }}">{{ doc.title or doc.name }}</a>
-</div>
-<!-- this is a sample default list template -->
