--- community/community/web_form/update_profile/update_profile.js
@@ -1,5 +1,0 @@
-frappe.ready(function() {
-	frappe.web_form.after_save = () => {
-		window.location.href = frappe.web_form.get_value("username")
-	}
-})
