--- community/community/web_form/update_profile/update_profile.json
@@ -1,100 +1,0 @@
-{
- "accept_payment": 0,
- "allow_comments": 0,
- "allow_delete": 0,
- "allow_edit": 1,
- "allow_incomplete": 0,
- "allow_multiple": 0,
- "allow_print": 0,
- "amount": 0.0,
- "amount_based_on_field": 0,
- "apply_document_permissions": 0,
- "breadcrumbs": "",
- "button_label": "Save",
- "creation": "2021-03-09 17:34:03.394301",
- "doc_type": "Community Member",
- "docstatus": 0,
- "doctype": "Web Form",
- "idx": 0,
- "is_standard": 1,
- "login_required": 1,
- "max_attachment_size": 0,
- "modified": "2021-03-22 12:04:22.571655",
- "modified_by": "Administrator",
- "module": "Community",
- "name": "update-profile",
- "owner": "Administrator",
- "payment_button_label": "Buy Now",
- "published": 1,
- "route": "edit-profile",
- "route_to_success_link": 0,
- "show_attachments": 0,
- "show_in_grid": 0,
- "show_sidebar": 0,
- "sidebar_items": [],
- "success_message": "Profile updated successfully.",
- "success_url": "/",
- "title": "Update Profile",
- "web_form_fields": [
-  {
-   "allow_read_on_all_link_options": 0,
-   "fieldname": "full_name",
-   "fieldtype": "Data",
-   "hidden": 0,
-   "label": "Full Name",
-   "max_length": 0,
-   "max_value": 0,
-   "read_only": 0,
-   "reqd": 1,
-   "show_in_filter": 0
-  },
-  {
-   "allow_read_on_all_link_options": 0,
-   "fieldname": "username",
-   "fieldtype": "Data",
-   "hidden": 0,
-   "label": "User Name",
-   "max_length": 0,
-   "max_value": 0,
-   "read_only": 0,
-   "reqd": 1,
-   "show_in_filter": 0
-  },
-  {
-   "allow_read_on_all_link_options": 0,
-   "fieldname": "short_intro",
-   "fieldtype": "Data",
-   "hidden": 0,
-   "label": "Short Intro",
-   "max_length": 0,
-   "max_value": 0,
-   "read_only": 0,
-   "reqd": 0,
-   "show_in_filter": 0
-  },
-  {
-   "allow_read_on_all_link_options": 0,
-   "fieldname": "bio",
-   "fieldtype": "Data",
-   "hidden": 0,
-   "label": "Bio",
-   "max_length": 0,
-   "max_value": 0,
-   "read_only": 0,
-   "reqd": 0,
-   "show_in_filter": 0
-  },
-  {
-   "allow_read_on_all_link_options": 0,
-   "fieldname": "photo",
-   "fieldtype": "Attach Image",
-   "hidden": 0,
-   "label": "Photo",
-   "max_length": 0,
-   "max_value": 0,
-   "read_only": 0,
-   "reqd": 0,
-   "show_in_filter": 0
-  }
- ]
-}
