--- community/community/web_form/update_profile/update_profile.py
@@ -1,7 +1,0 @@
-from __future__ import unicode_literals
-
-import frappe
-
-def get_context(context):
-	# do your magic here
-	pass
