+++ community/patches.txt
@@ -2,3 +2,4 @@ community.patches.set_email_preferences
 community.patches.change_name_for_community_members
 community.patches.save_abbr_for_community_members
 community.patches.create_mentor_request_email_templates
+community.patches.replace_member_with_user_in_batch_membership
