+++ community/www/courses/utils.py
@@ -37,16 +37,6 @@ def get_course(name):
     except frappe.DoesNotExistError:
         return
 
-def get_instructor(owner):
-    instructor = frappe._dict()
-    try:
-        instructor = frappe.get_doc("Community Member", {"email": owner})
-    except frappe.DoesNotExistError:
-        instructor.full_name = owner
-        instructor.abbr = ("").join([ s[0] for s in owner.split() ])
-    instructor.course_count = len(frappe.get_all("LMS Course", {"owner": owner}))
-    return instructor
-
 def get_batch_members(batch):
     members = []
     memberships = frappe.get_all("LMS Batch Membership", {"batch": batch}, ["member", "member_type"])
